Separación y corrección de funciones de creación de sesiones
